backend/requirements.txt
python-dotenv
pymysql
